Handle non-georeferenced raster data
This adds support for non-georeferenced raster data in the form of PNG and
JPG files. We assume that the labels are still in the form of GeoJSON files
except that the CRS is just pixel col/row. To do this, we still use Rasterio to
read the files (since they may not fit into RAM) and use an IdentityCRSTransformer
which uses the identity function to map from pixel to map coords.